requirements.txt
# Core dependencies
requests>=2.28.0
python-dotenv>=1.0.0
loguru>=0.7.0
websocket-client>=1.5.0
aiohttp>=3.8.0

# Optional dependencies
pandas>=2.0.0  # For data analysis
numpy>=1.24.0  # For numerical operations
matplotlib>=3.7.0  # For plotting
pytest>=7.0.0  # For testing

# API Integration
ccxt==3.0.37

# Data Analysis
pandas==1.5.3
numpy==1.24.3

# Technical Analysis
pandas-ta==0.3.14b0
# Note: TA-Lib requires separate installation of C library
# Instructions: https://github.com/mrjbq7/ta-lib#dependencies

# Visualization
seaborn==0.12.2

# Backtesting
backtrader==1.9.76.123

# Testing
pytest==7.3.1

# Utilities
tqdm==4.65.0
pytz==2023.3 

dash==2.11.1
dash-bootstrap-components==1.4.2
plotly==5.14.1
pandas>=1.5.0
numpy>=1.24.0
flask>=2.2.0
loguru>=0.7.0
python-dateutil>=2.8.2 